Add more requirements for when deleting a unit is allowed
- Unit must belong to the currently active game
- Unit must belong to the player whose turn it is
- Unit must have at least 1 remaining movement point

+++ routes/unit.js
@@ -1,5 +1,6 @@
 const {
   Player,
+  Session,
   Unit,
 } = require('./import');
 
@@ -8,12 +9,23 @@ module.exports = {
 };
 
 async function deleteOneUnit(req, res) {
-  Unit.findById(req.params.id, (error, unit) => {
+  const currentGameId = Session.getCurrentGameId(req);
+  if (!currentGameId) {
+    return res.status(401).send(null);
+  }
+  Unit.findOne({_id: req.params.id, game: currentGameId}, async (error, unit) => {
     if (error) {
       return res.status(500).send(error);
     }
     if (!unit) {
       return res.status(404).send(null);
+    }
+    if (unit.movesRemaining === 0) {
+      return res.status(401).send(null);
+    }
+    const belongsToActiveUser = await Unit.belongsToActiveUser(unit);
+    if (!belongsToActiveUser) {
+      return res.status(401).send(null);
     }
     Unit.deleteOne({_id: req.params.id}, (error, value) => {
       if (error) {
